backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyupbit
pandas
requests
